feat: move HTTPClientEvent constants to src/Constants
